pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "openpyxl>=3.1.5",
    "sqlalchemy>=2.0.42",
    "werkzeug>=3.1.3",
    "anthropic>=0.60.0",
    "google-genai>=1.28.0",
    "twilio>=9.7.0",
    "trafilatura>=2.0.0",
    "langsmith>=0.4.9",
    "langchain>=0.3.27",
    "requests>=2.32.4",
    "sift-stack-py>=0.8.1",
]
